Update Stash_FI.txt
By me, a trained professional EN-FI translator. Opted to emulate the rushed and colloquial tone of EN map annotations.

Was unclear on how to interpret "fucked up" in the context of a map annotation. Did the person make a mistake here, maybe get bitten? Or is the location broken down, ravaged by zombies? Or maybe they saw disturbing things here, aka the place is full of zombies? Opted to interpret "fucked up" as confessing to a mistake, and "all kinds of fucked up" to imply a large number of zombies, but I've been led to believe that map annotations are not just storytelling but mechanically connected to unique spawns etc, so can opt for another translation for those if given more info on context.